Avoid Inferring the Response MIME Type with helmet.noSniff()
As a reminder, this project is being built upon the following starter project on Replit, or cloned from GitHub. 

Browsers can use content or MIME sniffing to override response Content-Type headers to guess and process the data using an implicit content type. While this can be convenient in some scenarios, it can also lead to some dangerous attacks. This middleware sets the X-Content-Type-Options header to nosniff, instructing the browser to not bypass the provided Content-Type.

Use the helmet.noSniff() method on your server.