Disable DNS Prefetching with helmet.dnsPrefetchControl()
As a reminder, this project is being built upon the following starter project on Replit, or cloned from GitHub.

To improve performance, most browsers prefetch DNS records for the links in a page. In that way the destination ip is already known when the user clicks on a link. This may lead to over-use of the DNS service (if you own a big website, visited by millions people…), privacy issues (one eavesdropper could infer that you are on a certain page), or page statistics alteration (some links may appear visited even if they are not). If you have high security needs you can disable DNS prefetching, at the cost of a performance penalty.

Use the helmet.dnsPrefetchControl() method on your server.